+++ pdns/nameserver.cc
@@ -254,7 +254,6 @@ void UDPNameserver::send(DNSPacket *p)
   g_rs.submitResponse(p->qtype.getCode(), buffer.length(), true);
 
   struct msghdr msgh;
-  struct cmsghdr *cmsg;
   struct iovec iov;
   char cbuf[256];
 
@@ -285,53 +284,7 @@ void UDPNameserver::send(DNSPacket *p)
   msgh.msg_namelen = p->d_remote.getSocklen();
 
   if(p->d_anyLocal) {
-    if(p->d_anyLocal->sin4.sin_family == AF_INET6) {
-      struct in6_pktinfo *pkt;
-          
-      msgh.msg_control = cbuf;
-      msgh.msg_controllen = CMSG_SPACE(sizeof(*pkt));
-                  
-      cmsg = CMSG_FIRSTHDR(&msgh);
-      cmsg->cmsg_level = IPPROTO_IPV6;
-      cmsg->cmsg_type = IPV6_PKTINFO;
-      cmsg->cmsg_len = CMSG_LEN(sizeof(*pkt));
-                                  
-      pkt = (struct in6_pktinfo *) CMSG_DATA(cmsg);
-      memset(pkt, 0, sizeof(*pkt));
-      pkt->ipi6_addr = p->d_anyLocal->sin6.sin6_addr;
-      msgh.msg_controllen = cmsg->cmsg_len; // makes valgrind happy and is slightly better style
-    }
-    else {
-#ifdef IP_PKTINFO
-      struct in_pktinfo *pkt;
-      msgh.msg_control = cbuf;
-      msgh.msg_controllen = CMSG_SPACE(sizeof(*pkt));
-
-      cmsg = CMSG_FIRSTHDR(&msgh);
-      cmsg->cmsg_level = IPPROTO_IP;
-      cmsg->cmsg_type = IP_PKTINFO;
-      cmsg->cmsg_len = CMSG_LEN(sizeof(*pkt));
-
-      pkt = (struct in_pktinfo *) CMSG_DATA(cmsg);
-      memset(pkt, 0, sizeof(*pkt));
-      pkt->ipi_spec_dst = p->d_anyLocal->sin4.sin_addr;
-#endif
-#ifdef IP_SENDSRCADDR
-      struct in_addr *in;
-    
-      msgh.msg_control = cbuf;
-      msgh.msg_controllen = CMSG_SPACE(sizeof(*in));
-            
-      cmsg = CMSG_FIRSTHDR(&msgh);
-      cmsg->cmsg_level = IPPROTO_IP;
-      cmsg->cmsg_type = IP_SENDSRCADDR;
-      cmsg->cmsg_len = CMSG_LEN(sizeof(*in));
-                            
-      in = (struct in_addr *) CMSG_DATA(cmsg);
-      *in = p->d_anyLocal->sin4.sin_addr;
-#endif
-      msgh.msg_controllen = cmsg->cmsg_len;
-    }
+    addCMsgSrcAddr(&msgh, cbuf, p->d_anyLocal.get_ptr());
   }
   DLOG(L<<Logger::Notice<<"Sending a packet to "<< p->getRemote() <<" ("<< buffer.length()<<" octets)"<<endl);
   if(buffer.length() > p->getMaxReplyLen()) {
